pyproject.toml
[tool.poetry]
name = "todolist"
version = "0.1.0"
description = ""
authors = ["ZooBycH <mz88@list.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.4"
django-environ = "^0.9.0"
psycopg2-binary = "^2.9.5"
djangorestframework = "^3.14.0"
social-auth-app-django = "^5.0.0"
django-cors-headers = "^3.13.0"
django-filter = "^22.1"
marshmallow-dataclass = "^8.5.10"
requests = "^2.28.1"
pytest-django = "^4.5.2"
pytest-factoryboy = "^2.5.1"
drf-spectacular = "^0.25.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
